A várnak az alapkövei

+++ SFML/player.cpp
@@ -164,7 +164,7 @@ int player::getScore()
 }
 int player::getWeapon()
 {
-    float result = -1;
+    int result = -1;
     my_mutex.lock();
     result = weapon;
     my_mutex.unlock();
